splits Molr Web interfaces
