+++ client/src/main/java/cern/molr/client/impl/MissionExecutionServiceImpl.java
@@ -6,14 +6,11 @@ package cern.molr.client.impl;
 
 import cern.molr.client.api.ClientMissionController;
 import cern.molr.client.api.MissionExecutionService;
+import cern.molr.client.api.MolrClientToServer;
 import cern.molr.commons.api.request.MissionCommand;
 import cern.molr.commons.api.response.CommandResponse;
 import cern.molr.commons.api.response.MissionEvent;
 import cern.molr.commons.api.response.MissionState;
-import cern.molr.commons.api.web.MolrWebClient;
-import cern.molr.commons.api.web.MolrWebSocketClient;
-import cern.molr.commons.impl.web.MolrWebClientImpl;
-import cern.molr.commons.impl.web.MolrWebSocketClientImpl;
 import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,8 +28,7 @@ public class MissionExecutionServiceImpl implements MissionExecutionService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MissionExecutionServiceImpl.class);
 
-    private MolrWebClient client;
-    private MolrWebSocketClient clientSocket;
+    private MolrClientToServer client;
 
     /**
      * This constructor searches for MolR address (host and port) in "config.properties"
@@ -70,13 +66,11 @@ public class MissionExecutionServiceImpl implements MissionExecutionService {
                     "values", error);
         }
 
-        client = new MolrWebClientImpl(host, port);
-        clientSocket = new MolrWebSocketClientImpl(host, port);
+        client = new MolrClientToServerImpl(host, port);
     }
 
     public MissionExecutionServiceImpl(String host, int port) {
-        client = new MolrWebClientImpl(host, port);
-        clientSocket = new MolrWebSocketClientImpl(host, port);
+        client = new MolrClientToServerImpl(host, port);
     }
 
     @Override
@@ -85,17 +79,17 @@ public class MissionExecutionServiceImpl implements MissionExecutionService {
         return client.instantiate(missionName, missionArguments, missionId -> new ClientMissionController() {
             @Override
             public Publisher<MissionEvent> getEventsStream() {
-                return clientSocket.getEventsStream(missionName, missionId);
+                return client.getEventsStream(missionName, missionId);
             }
 
             @Override
             public Publisher<MissionState> getStatesStream() {
-                return clientSocket.getStatesStream(missionName, missionId);
+                return client.getStatesStream(missionName, missionId);
             }
 
             @Override
             public Publisher<CommandResponse> instruct(MissionCommand command) {
-                return clientSocket.instruct(missionName, missionId, command);
+                return client.instruct(missionName, missionId, command);
             }
         });
     }
