+++ server/src/main/java/cern/molr/server/impl/RemoteMoleSupervisorImpl.java
@@ -5,10 +5,8 @@ import cern.molr.commons.api.request.client.ServerInstantiationRequest;
 import cern.molr.commons.api.response.CommandResponse;
 import cern.molr.commons.api.response.MissionEvent;
 import cern.molr.commons.api.response.SupervisorState;
-import cern.molr.commons.api.web.MolrWebClient;
-import cern.molr.commons.api.web.MolrWebSocketClient;
-import cern.molr.commons.impl.web.MolrWebClientImpl;
-import cern.molr.commons.impl.web.MolrWebSocketClientImpl;
+
+import cern.molr.server.api.MolrServerToSupervisor;
 import cern.molr.server.api.RemoteMoleSupervisor;
 import org.reactivestreams.Publisher;
 
@@ -21,23 +19,21 @@ import java.util.Optional;
  */
 public class RemoteMoleSupervisorImpl implements RemoteMoleSupervisor {
 
-    protected MolrWebClient client;
-    private MolrWebSocketClient socketClient;
+    protected MolrServerToSupervisor client;
 
     public RemoteMoleSupervisorImpl(String host, int port) {
-        this.socketClient = new MolrWebSocketClientImpl(host, port);
-        this.client = new MolrWebClientImpl(host, port);
+        this.client = new MolrServerToSupervisorImpl(host, port);
     }
 
     @Override
     public <I> Publisher<MissionEvent> instantiate(ServerInstantiationRequest<I> serverRequest, String
             missionId) {
-        return socketClient.instantiate(serverRequest.getMissionName(), missionId, serverRequest.getMissionArguments());
+        return client.instantiate(serverRequest.getMissionName(), missionId, serverRequest.getMissionArguments());
     }
 
     @Override
     public Publisher<CommandResponse> instruct(MissionCommandRequest commandRequest) {
-        return socketClient.instruct("unknown", commandRequest.getMissionId(), commandRequest.getCommand());
+        return client.instruct("unknown", commandRequest.getMissionId(), commandRequest.getCommand());
     }
 
     @Override
