+++ test/src/test/java/cern/molr/test/supervisor/SupervisorTest.java
@@ -10,7 +10,7 @@ import cern.molr.commons.api.web.SimpleSubscriber;
 import cern.molr.commons.commands.MissionControlCommand;
 import cern.molr.commons.events.*;
 import cern.molr.commons.impl.mission.AnnotatedMissionMaterializer;
-import cern.molr.commons.impl.web.MolrWebSocketClientImpl;
+import cern.molr.commons.web.WebFluxWebSocketClient;
 import cern.molr.commons.web.MolrConfig;
 import cern.molr.sample.mission.Fibonacci;
 import cern.molr.supervisor.RemoteSupervisorMain;
@@ -171,7 +171,7 @@ public class SupervisorTest {
 
         InstantiationRequest<Integer> request =
                 new InstantiationRequest<>("1", Fibonacci.class.getName(), 42);
-        MolrWebSocketClientImpl client = new MolrWebSocketClientImpl("http://localhost", 8080);
+        WebFluxWebSocketClient client = new WebFluxWebSocketClient("http://localhost", 8080);
 
         client.receiveFlux(MolrConfig.INSTANTIATE_PATH, MissionEvent.class, request)
                 .doOnError(Throwable::printStackTrace)
